01_Primeros_pasos_Python/07_Curso_Django_Rest_Framework/practicas/patients/views_example_LISTAPIview.py
from .serializers import PatientSerializer
from .models import Patient

from rest_framework.generics import ListAPIView, CreateAPIView, RetrieveAPIView, UpdateAPIView,  DestroyAPIView, RetrieveUpdateDestroyAPIView  

class ListPatientsView(ListAPIView, CreateAPIView):
    allowe_methods = ["GET", "POST"]
    serializer_class = PatientSerializer
    queryset = Patient.objects.all()

class DetailPatientsView(RetrieveUpdateDestroyAPIView):
    allowe_methods = ["GET", "PUT", "DELETE"]
    serializer_class = PatientSerializer
    queryset = Patient.objects.all()