01_Primeros_pasos_Python/07_Curso_Django_Rest_Framework/practicas/requirements.txt
Django==5.2.1
djangorestframework==3.16.0
drf-spectacular==0.28.0
